make all plug-api files usable without an import_map
