+++ common/common_system.ts
@@ -1,4 +1,5 @@
-import { AppCommand, CommandHook } from "./hooks/command.ts";
+import { AppCommand } from "$lib/command.ts";
+import { CommandHook } from "./hooks/command.ts";
 import { PlugNamespaceHook } from "$common/hooks/plug_namespace.ts";
 import { SilverBulletHooks } from "./manifest.ts";
 import { buildQueryFunctions } from "./query_functions.ts";
