+++ common/space_script.ts
@@ -1,7 +1,7 @@
 import { System } from "../lib/plugos/system.ts";
 import { ParseTree } from "../plug-api/lib/tree.ts";
 import { ScriptObject } from "../plugs/index/script.ts";
-import { AppCommand, CommandDef } from "./hooks/command.ts";
+import { AppCommand, CommandDef } from "$lib/command.ts";
 import { Intl, Temporal, toTemporalInstant } from "@js-temporal/polyfill";
 
 // @ts-ignore: temporal polygifill
