+++ lib/command.ts
@@ -1,0 +1,28 @@
+export type CommandDef = {
+  name: string;
+
+  contexts?: string[];
+
+  // Default 0, higher is higher priority = higher in the list
+  priority?: number;
+
+  // Bind to keyboard shortcut
+  key?: string;
+  mac?: string;
+
+  hide?: boolean;
+  requireMode?: "rw" | "ro";
+};
+
+export type AppCommand = {
+  command: CommandDef;
+  run: (args?: any[]) => Promise<void>;
+};
+
+export type CommandHookT = {
+  command?: CommandDef;
+};
+
+export type CommandHookEvents = {
+  commandsUpdated(commandMap: Map<string, AppCommand>): void;
+};
