+++ plug-api/deps.ts
@@ -1,0 +1,1 @@
+export * as YAML from "https://esm.sh/js-yaml@4.1.0";
