+++ plug-api/lib/tags.ts
@@ -1,4 +1,4 @@
-import { FrontMatter } from "$sb/lib/frontmatter.ts";
+import { FrontMatter } from "./frontmatter.ts";
 import { ObjectValue } from "../types.ts";
 
 export function updateITags<T>(obj: ObjectValue<T>, frontmatter: FrontMatter) {
