+++ plug-api/syscalls/datastore.ts
@@ -1,4 +1,4 @@
-import { syscall } from "$sb/syscall.ts";
+import { syscall } from "../syscall.ts";
 import { KV, KvKey, KvQuery } from "../types.ts";
 
 export function set(key: KvKey, value: any): Promise<void> {
