+++ plug-api/syscalls/system.ts
@@ -1,4 +1,4 @@
-import type { CommandDef } from "$common/hooks/command.ts";
+import type { CommandDef } from "../../lib/command.ts";
 import type { SyscallMeta } from "../types.ts";
 import type { ParseTree } from "../lib/tree.ts";
 import { syscall } from "../syscall.ts";
