+++ type/web.ts
@@ -1,43 +1,15 @@
 import { Manifest } from "$common/manifest.ts";
 import { PageMeta } from "../plug-api/types.ts";
-import { AppCommand } from "$common/hooks/command.ts";
+import { AppCommand } from "../lib/command.ts";
 import { defaultSettings } from "$common/settings.ts";
-
-// Used by FilterBox
-export type FilterOption = {
-  name: string;
-  description?: string;
-  orderId?: number;
-  hint?: string;
-} & Record<string, any>;
-
-export type Notification = {
-  id: number;
-  message: string;
-  type: "info" | "error";
-  date: Date;
-};
-
-export type PanelMode = number;
-
-export type Shortcut = {
-  key?: string;
-  mac?: string;
-  priority?: number;
-  command: string;
-};
-
-export type ActionButton = {
-  icon: string;
-  description?: string;
-  command: string;
-  args?: any[];
-  mobile?: boolean;
-};
-
-export type EmojiConfig = {
-  aliases: string[][];
-};
+import {
+  ActionButton,
+  EmojiConfig,
+  FilterOption,
+  Notification,
+  PanelMode,
+  Shortcut,
+} from "$lib/web.ts";
 
 export type BuiltinSettings = {
   indexPage: string;
