+++ web/components/page_navigator.tsx
@@ -1,5 +1,5 @@
 import { FilterList } from "./filter.tsx";
-import { FilterOption } from "../../type/web.ts";
+import { FilterOption } from "$lib/web.ts";
 import { CompletionContext, CompletionResult } from "../deps.ts";
 import { PageMeta } from "../../plug-api/types.ts";
 import { isFederationPath } from "$sb/lib/resolve.ts";
