+++ web/fuse_search.ts
@@ -1,6 +1,6 @@
 // @deno-types="https://deno.land/x/fuse@v6.4.1/dist/fuse.d.ts"
 import Fuse from "fuse";
-import { FilterOption } from "../type/web.ts";
+import { FilterOption } from "$lib/web.ts";
 
 type FuseOption = FilterOption & {
   baseName: string;
